backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
#torch==2.3.1
#transformers==4.42.3
scikit-learn==1.5.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytest==8.2.2
sqlalchemy==2.0.20
